containers/autosklearn/requirements.txt
auto-sklearn==0.8
category_encoders
mlxtend
